gf_factory_data_rom: Created a new dyn block DYN_BLOCK_MX8M_RAM_SIZE_V2 to expand total_size variable to 64-bit

+++ arch/arm/include/asm/arch-imx8m/ddr.h
@@ -683,7 +683,7 @@ struct __attribute__((__packed__)) dram_fsp_msg {
 };
 
 struct __attribute__((__packed__)) dram_timing_info {
-	unsigned int total_size;
+	u64 total_size;
 	/* umctl2 config */
 	struct dram_cfg_param *ddrc_cfg;
 	unsigned int ddrc_cfg_num;
